commit multi-task learning

+++ prediction_models/tile_concat_wy/train/trainSeg.py
@@ -42,7 +42,7 @@ class Train(object):
                 labels = labels.view(-1, h, w).long().cuda()
                 loss1 = criterion[0](outputs['out'], labels)
                 loss2 = criterion[0](outputs['aux'], labels)
-                loss3 = criterion[1](outputs['isup_grade'], grade.cuda())
+                loss3 = criterion[1](outputs['isup_grade'], grade.float().cuda())
                 loss = loss1 + 0.4 * loss2 + loss3
                 train_loss.append(loss.item())
                 loss.backward()
